requirements.txt
asgiref==3.7.2
Django==5.0.1
django-appconf==1.0.6
django-compressor==4.4
django-debug-toolbar==4.3.0
django-htmx==1.17.3
django-modeltranslation==0.18.11
django-pandas==0.6.7
numpy==1.26.4
pandas==2.2.1
pillow==10.2.0
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
rcssmin==1.1.1
rjsmin==1.2.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2024.1
